requirements.txt
quart
Quart-Discord
tortoise-orm==0.17.8
asyncpg
discord